refactor(config): isolate world storage
